+++ src/lib/builder.dart
@@ -1,7 +1,7 @@
 library orm_m8_generator;
 
 import 'package:build/build.dart';
-import 'package:flutter_sqlite_m8_generator/m8_builder.dart';
+import 'package:f_orm_m8_sqlite/m8_builder.dart';
 
 Builder ormBuilder() {
   return M8Builder.withWrapper(