+++ src/lib/generator/core.dart
@@ -1,5 +1,5 @@
 library orm_m8_generator.core;
 
-export 'package:flutter_sqlite_m8_generator/generator/emitted_entity.dart';
-export 'package:flutter_sqlite_m8_generator/generator/entity_writer.dart';
-export 'package:flutter_sqlite_m8_generator/generator/model_parser.dart';
+export 'package:f_orm_m8_sqlite/generator/emitted_entity.dart';
+export 'package:f_orm_m8_sqlite/generator/entity_writer.dart';
+export 'package:f_orm_m8_sqlite/generator/model_parser.dart';