+++ src/lib/generator/utils/entity_attribute.dart
@@ -1,6 +1,6 @@
 import 'package:f_orm_m8/f_orm_m8.dart';
-import 'package:flutter_sqlite_m8_generator/generator/utils/attribute_metadata_parser.dart';
-import 'package:flutter_sqlite_m8_generator/generator/utils/type_mapper.dart';
+import 'package:f_orm_m8_sqlite/generator/utils/attribute_metadata_parser.dart';
+import 'package:f_orm_m8_sqlite/generator/utils/type_mapper.dart';
 
 class EntityAttribute {
   final String modelTypeName;