+++ src/lib/generator/utils/utils.dart
@@ -1,6 +1,6 @@
 library orm_m8_generator.core.utils;
 
-export 'package:flutter_sqlite_m8_generator/generator/utils/type_utils.dart';
-export 'package:flutter_sqlite_m8_generator/generator/utils/db_entity_type_enum.dart';
-export 'package:flutter_sqlite_m8_generator/generator/utils/entity_attribute.dart';
-export 'package:flutter_sqlite_m8_generator/generator/utils/format_utils.dart';
+export 'package:f_orm_m8_sqlite/generator/utils/type_utils.dart';
+export 'package:f_orm_m8_sqlite/generator/utils/db_entity_type_enum.dart';
+export 'package:f_orm_m8_sqlite/generator/utils/entity_attribute.dart';
+export 'package:f_orm_m8_sqlite/generator/utils/format_utils.dart';