+++ src/lib/generator/writers/entity_writer_factory.dart
@@ -1,8 +1,8 @@
-import 'package:flutter_sqlite_m8_generator/generator/core.dart';
-import 'package:flutter_sqlite_m8_generator/generator/utils/utils.dart';
-import 'package:flutter_sqlite_m8_generator/generator/writers/account_entity_writer.dart';
-import 'package:flutter_sqlite_m8_generator/generator/writers/account_related_entity_writer.dart';
-import 'package:flutter_sqlite_m8_generator/generator/writers/independent_entity_writer.dart';
+import 'package:f_orm_m8_sqlite/generator/core.dart';
+import 'package:f_orm_m8_sqlite/generator/utils/utils.dart';
+import 'package:f_orm_m8_sqlite/generator/writers/account_entity_writer.dart';
+import 'package:f_orm_m8_sqlite/generator/writers/account_related_entity_writer.dart';
+import 'package:f_orm_m8_sqlite/generator/writers/independent_entity_writer.dart';
 
 class EntityWriterFactory {
   EntityWriter getWriter(EmittedEntity emittedEntity) {