+++ src/lib/generator/writers/sql_definition_writer.dart
@@ -1,6 +1,6 @@
 import 'package:f_orm_m8/f_orm_m8.dart';
-import 'package:flutter_sqlite_m8_generator/generator/core.dart';
-import 'package:flutter_sqlite_m8_generator/generator/utils/utils.dart';
+import 'package:f_orm_m8_sqlite/generator/core.dart';
+import 'package:f_orm_m8_sqlite/generator/utils/utils.dart';
 
 class SqlDefinitionWriter with Spacers {
   final EmittedEntity emittedEntity;