+++ src/test/account_related_composite_entity_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
-import 'package:flutter_sqlite_m8_generator/generator/core.dart';
-import 'package:flutter_sqlite_m8_generator/generator/writers/entity_writer_factory.dart';
+import 'package:f_orm_m8_sqlite/generator/core.dart';
+import 'package:f_orm_m8_sqlite/generator/writers/entity_writer_factory.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:test/test.dart';
 import 'utils/test_annotation_utils.dart';