+++ src/test/column_metadata_level_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_sqlite_m8_generator/generator/core.dart';
+import 'package:f_orm_m8_sqlite/generator/core.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:test/test.dart';
 import 'utils/test_annotation_utils.dart';